[ALPS07792024] sapu: support 64-bit ffa handle
[Detail]
1. Replace 32-bit secure handle with 64-bit ffa handle
2. Remove platdata - platname

MTK-Commit-Id: 7f195871839521cab0ea5866f21e24ef221f2eab

CR-Id: ALPS07792024
Feature: np_security

+++ drivers/misc/mediatek/apusys/sapu/sapu_driver.h
@@ -13,7 +13,7 @@
 #define APUSYS_POWER_CONTROL	_IOWR(APUSYS_SAPU_IOC_MAGIC, 2, u32)
 
 struct sapu_ha_tranfer {
-	uint32_t handle;
+	uint64_t handle;
 	dma_addr_t dma_addr;
 	uint64_t model_hd_ha;
 };
